Create RemoteTranslation structure to allow to resources without translate to be sending to MicrosoftTranslationClient
RemoteTranslation model: Every RemoteTranslate will be a DelayedJob request to translatorAPI. We need that RemoteTranslation model to be polymorphic to use on proposals, debates, comments and legislations processes.

RemoteTranslation controller: It receive submit from views and create RemoteTranslations when resource has not translations enqueued.

RemoteTraslation caller: This module is in charge of parsing each RemoteTranslation, extracting the associated resource and preparing its fields to be sent to the MicrosoftTranslateClient.