Allow create translations without length validation
We have changed validate_translation method on Globalize concern.
The objective is skip length validations when locale is distinct to default_locale.

First we force apply :length validations when locale is equal to default_locale. After we reject :length from options and apply rest of validation options only when we have more than 1 options.
Ej: options = { length: "maximum: 10" }
When reject :length option in this example, options is equal to a empty hash and we cant execute validations.