Create new CustomCensusAPI
Create a new CustomCensusAPI with the same functionality as
the old CensusAPI.

In the new CustomCensusAPI both the request and the
processing of the response are made according to the
parameters defined in the "Remote Census Connection"
configuration page.

Same as old Census API we only consider document_type and
document_number on API request.