Require depencies for CommentTree
Eventhoug the specs run fine, when running the application in production webserver we are getting an exception as Rails is not finding the `CommentTree` class from lib directory. It's being searched into CommentableAction module only.

By requiring this depedency explicitly in commentable_action concern module the `CommentTree`class is loaded correctly.