Do not serialize form if already serialized
This commit covers when a user make some changes and reloads the page. Without
this patch form will be serialized again but this time serialization process
will consider latest changes introduced by the user and not persisted yet as
as the original content and application will not be able to warn the user about
changes pending to be persisted. I was not able to reproduce this situation
within a  test using `page.refresh`, not sure about the reason of this behavior on
test environment.