Prevent votes on annotation comment replies
In some cases, it's not clear for users whether they're supporting the
comment made to the annotation or the reply itself.

Since, by default, when users cannot vote for a comment, we show a link
to signing in, we're hiding these links through CSS so we don't have to
add a custom view. We're also changing the Abilities::Common model so
it's impossible to vote for a comment reply anyway.