Use default host in confirmation email
I'm not sure why this was changed in commits a969aa35d4 and 41457dd16c.
For now I'm changing it back because the change was causing a failure in
one test because `ENV["SERVER_NAME"]` didn't exist.

I'm also removing the `protocol: https` part, since on production
environments we use `force_ssl` and so this part isn't needed.