Make Spanish the only available locale
Until now, users could access the page in a different language using the
`locale` parameter in the URL, which is not what we wanted.

This way we don't have to change anything related to the locale
switcher, since it's only rendered when there's more than one language
available.

Note we need to use a dynamic block in factories using Faker, since the
faker gem is loaded in the development environment and it depends on the
English locale being available. Using a dynamic block is actually what
recent versions of factory bot enforce and it's what recent versions of
Consul do.