Move login links back to their original position
Since we've added a block of top links, the login items are no longer
easily recognizable when they're at the top as well.

Besides, the HTML we were generating was invalid for two reasons.
First, it generated a duplicate `#responsive-menu` element. And, second,
it had a `<div>` tag as a direct child of a `<ul>` tag.