Configure URLs on header logos
The citizen participation logo must point to the citizen participation
URL, and the government logo must point to the government URL.

Since these URLs cannot be edited in the admin section, we're using
environment variables as Álvaro suggested.

With these changes, there's no link to the homepage. So we're adding a
link in the navigation. We could also do it with a custom block, but
IMHO in this case it makes more sense to have it in the source code
because we've always had a link to the homepage in the source code (it
was the logo).