Adjust European Development Program text styles
We're avoiding an accessibility issue, since we had a heading with no
content.

Note we're using `justify-content: flex-end` instead of the previous
`text-align: right`. The difference is subtle: when there are several
lines, with `text-align: right` the end of every line is aligned to the
right. But with `justify-content: flex-end`, only the end of the longest
line is aligned to the right. The given design uses the latter style,
and so do we.