Add content blocks to footer
As done in the design.

Note we're wrapping `<dt>` and `<dd>` items inside a `<div>` tag. This
is allowed since HTML 5.2.

We could remove the `<div>` tag and use a CSS multi-column layout.
However, we'd have to use a `dd { break-after: column }` rule, which is
currently only supported in 73% of the browsers.

A better option would be to use a grid layout, which currently works on
94.5% of the browsers. However, it doesn't work in Internet Explorer 11,
and we don't use it anywhere else in the code, so I've decided not to
introduce it here.

For future reference, the code with a grid layout would be something
like:

dl {
  @include breakpoint(large) {
    column-gap: 2 * rem-calc(map-get($grid-column-gutter, medium));
    display: grid;
    width: 75%;

    dt {
      grid-row-start: 1;
    }
  }
}