Unify ways of marking tests as broken
Broken tests were identified by using `:broken`, using `xscenario` or
commenting the test.

We're now using `:broken` everywhere. This way it's easy to find which
tests are broken.

The tests editing credentials, deleting an account or resetting the
password are marked as broken because those links point to the OpenAM
service.

The tests in signature sheets, officing, residence and management
verifications are broken due to the remote census; this feature has
changed in later Consul versions and we'll have to check how it works.

The tests in poll voters fail because there's some kind of issue with
the age or Census validations.

The Census behaving differently is also the cause Census API and Census
Caller tests are broken.

The budget votes specs are broken because now users need to have a valid
postal code at the time of voting (not just at the time of
verification).

The track tests and SMS verification tests fail because we don't use SMS
verification anymore.

Finally, the residence verification tests fail because we verify the
user when verifying the residence.