Change design of the main page header
Using colors similar to the ones in the Participación Ciudadana page.

Note we're making the font size increase as the window size increases.
That way we won't have ultra-long lines with more than 100 characters,
which are hard to read.

We're also aligning the text in the center on large screens, similar to
what Consul does. The main reason for it is we want to make the button
more prominent by putting it in the center of the screen, but if the
header description is short (like "Welcome"), on large screens we'll
have a short text aligned to the left with no apparent visual connection
to the button which is in the center. On small and medium screens we're
keeping the text aligned to the left, because it's easier to read.

Also note we're using a "trick" to make the title's background use the
whole width while keeping the title aligned with the rest of
Foundation's grid. We could achieve a similar effect by adding extra
`row` and `column` elements to the HTML, but that would make the HTML
far more complex.