Allow rendering translation fields separately.
Sometimes we want to render the translations fields separated from each other,
with not translatable fields or any other HTML component in between.
Unfortunately, this was not possible because the hidden fields `:_destroy` and
`:locale` were rendered for each `translatable_field` call resulting in
duplicated fields for each translation and invalid HTML. So now we ensure to
render each translation hidden_fields once per form.

In addition, the form fields for the same translation had a different index
when using more than one `translatable_fields` block. So now, the index we use
for translation fields is the position the translation locale occupies in the
application's array of available languages.

This is an example of what we want to wirte

```
    <%= f.translatable_fields do |translations_form| %>
        <%= translations_form.text_field :title %>
    <% end %>

     Some HTML here

    <%= f.translatable_fields do |translations_form| %>
        <%= translations_form.text_area :description %>
    <% end %>
```

The output for the same translation fields was:

```
    debate[translations_attributes][1][_destroy]
    debate[translations_attributes][1][locale]
    debate[translations_attributes][1][title]

    Some HTML here

    debate[translations_attributes][33][_destroy]
    debate[translations_attributes][33][locale]
    debate[translations_attributes][33][description]
```

After this patch, translation hidden fields are not duplicated and the index is
the same for the same translation fields:

```
    debate[translations_attributes][1][_destroy]
    debate[translations_attributes][1][locale]
    debate[translations_attributes][1][title]

    Some HTML here

    debate[translations_attributes][1][description]
```