Hide regular sign in form by default
This way users won't start filling in the form before realizing they're
supposed to click on the link to the OpenAM sign in form.

Note we're using the `<details>` element, which is currently supported
by 97% of the browsers. For the other 3% of users (mainly Internet
Explorer and Opera Mini users), the form will always be visible.

Also note our current version of Capybara/Selenium doesn't treat
`<summary>` elements as buttons. When we upgrade, we will be able to
write tests using `click_button "Or fill the following form:"`.