Fix map scrolling usability issue
Solve a usability issue when a user's scrolls on pages
with a map and additional content below. When the user
is scrolling, and the mouse goes over the map, the
window stops scrolling, and the map starts moving; this
behaviour annoys the users as they expect to keep
scrolling the page to the bottom. A similar problem
occurs when browsing the page on touchable devices
like phones or tablets.

We are disabling the mouse wheel map listener to keep
scrolling to the page bottom on desktop devices. We are
disabling the dragging and the tap features for mobile
devices, although as the user can not move the map, we
need arrow controls to allow users to change the map pan.

Credits for the mobile solution are for SimonEast [1].

[1] https://github.com/Leaflet/Leaflet/issues/5425#issuecomment-342055520