Remove expectation as the checkbox was removed in custom views
IMO if the user is still able to disable debates recommendations
from the debates recommendations section he should be able to
enable it from the user's account page. Other option is to remove
the "Hide recommendations" link from the debates
recommendations sections.

IMO if the user is still able to disable proposal recommendations
from the proposals recommendations section he should be able to
enable it from the user's account page. Other option is to remove
the "Hide recommendations" link from the proposals
recommendations sections, so the user cannot disable recommendations from anywhere.