Adding consistency to i18n task specs
The base_locale for i18n-tasks is :nl and for this language some translations
are missing.

This project contains 1.2.0 version where there are 490 missing keys for the
:nl language. As 48 keys have been added in the custom locales, we update
the missing keys spec to 442, to make sure we don't add any more.

As for the unused keys, until we clarify why 76 unused keys have been added,
we update the spec to make sure we don't add any more.