Chande te way to check permissions for comments controller create action
The `load_and_authorize_resource` method was checking the create action
permissions before creating comments but was not checking the
comment_as_administrator and comment_as_moderation permissions. This wasn't
an issue because every logged user was able to add comments everywhere so
the authorization was always returning true, but since we needed to change
the common users permissions and disable the ability to create comments on
proposals where voting is disabled or to review.

As a developer, when I read an ability definition like the following:

`can :comment_as_moderator, [Proposal, Comment, ...]`

I expect that the application allows the moderator to use the action create
on those classes. :thinking_face