Block interface to add new comment or reply
When a proposal is not enabled for voting, normal users will not see the
comment form or the reply link on existing comments. These users will
see a message instead, so the user knows the reason for comment
blocking.

Moderators see the form but will only be allowed to add comments when
the checkbox "Comment as moderator" is checked. If the moderator sends
the comment unchecking the checkbox, the action will be blocked by
abilities.

Administrators will behave mostly the same as moderators but using the
"Comment as administrator" checkbox instead.