Publish proposals after they've been reviewed
This way proposals which haven't been reviewed won't be shown in the
proposals index.

Another option would be to change the `published` scope as well as the
`published?` method in the proposal model, so they only return proposals
which have already been reviewed. We're not doing so because sometimes
we use the `published_at` attribute assuming proposals with that
attribute are published, and so we'd have to change/review the logic of
every place where we use that attribute.

Note we're forbidding the `publish` action in the proposals controller.
Even if we didn't have links/buttons to that action, users could still
access it by manually sending a PATCH request to the route to publish a
proposal.