Add recaptcha to registration user page
To achieve security certification we must replace the invisible
captcha with a visible captcha (reCaptcha gem).

In devise gem exist a section 'How To: Use Recaptcha with Devise' [1].
We apply the suggested changes.

This way we allow to validate the filled form fields and show the
errors even if the captcha is not validated.

As you can see in the original user registration controller we have the
following line:
> invisible_captcha only: [:create], honeypot: :address, scope: :user
In our custom controller we should skip this line. The problem is that
the invisible captcha does not provide an action name to reference so
we have to get the name some way. This is what we do with the
self.invisible_captcha_actions method. Then we can skip it by referencing
this method with: "skip_before_action(*invisible_captcha_actions)"

If there was a skip_invisible_captcha, all this would not be necessary.

[1] https://github.com/heartcombo/devise/wiki/How-To:-Use-Recaptcha-with-Devise