Do not change default `relative_url_root` when is not needed
Do not change default `relative_url_root` when it is not
needed

We were assigning `/` to the `relative_url_root` option
even when the environment variable CONSUL_RELATIVE_URL
was unset. IMO it's better to use the `relative_url_root`
option when the variable is unset.

For example, when running the application in
development without the `CONSUL_RELATIVE_URL`
most of the links were broken because the following
expression:

`(Rails.configuration.relative_url_root || "") + document.attachment.url`

which generates the following invalid URL

`//system/attachments/...`

Note the double slash at the beginning of the URL.