Calculate the pan controls offset from map dimensions
Using a fixed offset for all screen sizes does not
work nice. For example, when using mobile devices
(usually small), the default offset value (500px by
default) moves the map pan to a too far pan and
makes the user lose the reference point from the
previous pan.

Using the same offset for the vertical and horizontal
pan controls does not work nicely when the map does not
have an square shape.

To better control the vertical and horizontal offsets
and considering we show maps with all the shapes,
we will use different offsets for vertical and
horizontal panning calculated from the map
dimensions, so we can move the pan as much as
possible without losing the reference point.