Install Leaflet.GestureHandling plugin
Solve a usability issue when a user's scrolls on
pages with a map and additional content below of it.
When the user is scrolling, and the mouse goes over the
map, the window stops scrolling, and the map starts
moving; this behavior annoys the users as they expect to
keep scrolling the page to the bottom. A similar problem
occurs when browsing the page on touchable devices
like phones or tablets.

Google handles this problem with a plugin called Google
Maps Gesture Handling. We are installing a plugin that
adds the same behavior to Leaflet Maps.

Quoting the Leaflet.GestureHandling library README:
> This library brings the basic functionality of Google
Maps Gesture Handling into Leaflet.

1. Install Leaflet.GestureHandling plugin v1.2.1 files [1]
    into the vendor assets folder.
2. Configure CONSUL maps to enable the feature

[1] Check Github elmarquis/Leaflet.GestureHandling project v1.2.1