Integrate relative URL root with Paperclip
We were manually adding the root URL before every attachment, but we can
configure it with the `url:` paperclip option.

Note by default the value of the `path:` option is
`:rails_root/public:url`, but we don't want to add the relative URL to
the path, so we're adding this option to all paperclip attachments.

The code is based on ideas mentioned in the issue 889 in the Paperclip
repository.