Fix evaluation comment email on system emails
Currently with both seeds and dev_seeds, not only was this email not
displayed from the system emails section, but it also caused an error in
the application.

@email_to had an empty value and in the view we tried to access
@email_to.name which caused the error. We kept the same logic but
added the Administrator to make sure it always has a valid value. We add
the Adminsitrator and not the valuator, because the Administrator is always present in both seeds and dev seeds.