Remove omniauth-saml from omniauth form
We want the ominauth access via saml to be rendered in another section of the
session / registration login page. For this reason we remove the saml section and
will add it in a later commit in the right place.

Recover "or fill" text to make sense of this section again. Instead of overwriting
the text in the "custom" file, we reuse the "original". But since it has been
modified in past commits, we get back the text from Consul.