Force password complexity
Devise-security allows to define a validation for password complexity.

This validation was not being applied because the "self.inlcuded" method
from SecureValidatable module had been overwritten and only the
:current_equal_password_validation was being defined.

To keep the same logic on the :current_equal_password_validation,
just overwrite this method and not the self.included method.

Now when using the original self.included method, the validation defined
in password_regex is already applied.

The increase in complexity is that the passwords now need to contain
at least 1 uppercase character, 1 lowercase character and 1 number.
The password length has not been modified and is still defined in the
Devise initializer as "config.password_length = 8..72"

In this commit all the necessary modifications are made in the specs
so that now this new restriction is fulfilled.