Unify comment icon section on legislation annotation index
Apply new structure in the section that shows the comments icon together with the number of comments so that it is easier to unify them into one component.

In this case we make only the text clickable and not the icon as in the rest
of the application. We keep the color and text-decoration "for now".