Unify comment icon section on legislation questions
Remove unnecessary span class "debate-comments"

We take advantage of this commit to also unify the format of the date that appears
next to the comments with the rest of the application. The format that we
removed is being used on the same page in the "Participation phases" tab
(I guess that was the reason for putting it the same) but I think it makes more
sense to use the date format that is used in this kind section in the rest of the
application.