Force password with complexity required
Since commit 0b2bb35, we need all passwords must contain at least one digit
and at least one upper-case letter.

We were having some flakies because of the Devise.friendly_token that we use to
generate passwords sometimes it generated passwords that did not valid
because they did not contain any digit.

We updated the code to always add a random digit between 0 and 9, in a random
position of the generated token.

We have modified this line:
--  Devise.friendly_token[0, 20]
++ Devise.friendly_token.insert(rand(1..20), rand(0..9).to_s)

In Devise friendly_token method by default, length is 20 characters, so we can
avoid passing the parameter