Ignore LDAP roles not related to Consul
We were having an issue when the SAML response sent us roles in the
`isMemberOf` field which were not related to Consul and so the regular
expression to capture the name of the group returned `nil`. In this
case, the application was crashing.

So now we're handling this case by simply using the safe navigation
operator.