Sign SAML logout requests
For some reason, the CAS we're connecting to doesn't require
authorization requests to be signed, but it does require the signature
on logout requests. It doesn't make much sense, but there's nothing we
can do, so we're signing these requests.

Note that we must generate a valid certificate with its private key in
order to make the logout work properly.