Create new admin layout for only manage tenant configuration
We want only render account link and login items in the header.
And we want only render Multitenancy section in the admin sidebar.

In order not to add complexity and conditions to the existing layouts
and views for this specific case, I think the best solution is to
create a new layout so that we can define it more easily.

A new menu tenant component is also created to facilitate adding
new sections for this configuration in the future.