Fix alt text in Cabildo Tenerife logo
We were showing the org_name as the alt text, and we had another link
having the org_name as text next to it which pointed to a different URL.
That meant people using screen readers were hearing two consecutive
links with the same text and pointing to different URLs, which was
confusing.

So we're using the same text that appears in the image as alt text.