Fix crash deleting a geozone with no associations
The application was crashing since we added the relationship with audits
in commit d728bd6f6 because audits don't belong to a geozone but to an
auditable. The tests "Delete geozone with no associated elements" and
"#safe_to_destroy? is true when not linked to other models" were failing
because of this.