Fix schema.rb file
We accidentally added the main_link_text and main_link_url fields (which
will be added in Consul 1.4.0) in commit 900a1cc40. And the migration
introduced in commit 44839c141 wasn't in sync with the schema it
produces.

So now we get the same contents we get when running db:migrate and we
guarantee consistency in every database related to this project.