Do not render the consent banner in the test environment
As its interface overlaps with the application interface, causing
errors when interacting with elements behind the consent
banner layer.

As the cookies consent banner playing with cookies in system specs
is tricky, I decided to disable the consent banner for all the test
suite except in the specs testing the consent banner behaviour. In
those specs, we can mock the secret entry with the code below.

`allow(Layout::ConsentBannerComponent).to receive(:render).and_return(true)`