Fix tests failing because of the consent banner
The cookie consent banner is displayed on top of the rest of the page,
meaning it sometimes makes it impossible to interact with the elements
behind it.

In the cases of voting related content and commenting on existing
annotations, tests were failing because of that. For now we aren't
displaying the banner in these tests. Once we upgrade to version 1.4.0,
we probably won't need this hack in the related content tests because in
that version the accessibility of this section was improved and so we
don't need to use `hover` in the tests, which we think is the issue.

We've considered disabling the cookie banner in the test environment,
but since it will only affect one place once we upgrade to version
1.4.0, keeping it might be a good way to find places when the UX isn't
that great (which is the case with both related content and
annotations).

Co-Authored-By: Senén Rodero Rodríguez <senenrodero@gmail.com>