Update custom Proposal::Exporter to work with ML
In CabildoAbierto the Proposal::Exporter model had been added in the
custom folder. This model did not exist in version 1.3.1, so it  did not
depend on any other model.

Now with the update to 1.4.1 we have found that a ProposalExporter
model has been added for the Machine Learning feature and it is giving
us problems.

To keep both functionalities working we have added the dependency of
the custom Proposal::Exporter to the Proposal::Exporter. So both calls
either to export proposals to CSV (Cabildo development) or to export
proposals to json (machine learning development introduced in 1.4.1) can
continue to work correctly.

The custom model Proposal::Exporter has an initializer that has to receive
an Array, so we do the same thing that is done in machine learning
module to export the budget investments (which happens in the same
case).

The simplest solution if we do not want to enter in a change of names, is
to add an Array as default in initialize method.