Avoid DoubleLinkError with custom images
Quoting from Consul 1.5.0 release notes:

Due to an upgrade of the sprockets gem, if you're overwriting images
under the app/assets/images/ folder by adding an image with the same
name to the app/assets/images/custom/ folder, you might run into a
Sprockets::DoubleLinkError. As a workaround, you can delete or rename
the image under app/assets/images/ which has the conflicting name.

In our case, this only affects one image: `logo_email.png`. We're
renaming the file instead of deleting it in order to avoid conflicts if
new versions of Consul change it.