Add model concern to manage audit feature
Note we are monkey patching the globalize translation class as
it does not inherits from ApplicationRecord and therefore the
audit feature needs to be enabled explicitly. The same happens
with the Tagging model.

We excluded the following classes from auditing:

ActivePoll Activity Administrator AdminNotification Ahoy:Event Ckeditor::Asset
Lock MachineLearningInfo MachineLearningJob Manager MlSummaryComment Moderator
RemoteTranslation Visit

Note also that this version recovers the original implementation
of the `safe_to_destroy?` method from the Geozone model, as we should've done
this in the custom file.