Simplify footer and top links colors
The colors of the text in the footer and the top links are now
calculated based on the background color stored in the $footer and
$top-links variables, so we can use the variables instead of directly
changing the background color and the color of the links.

Since the $top-links variable is already $brand-secondary by default, we
don't even need to change this one.