Disable last_attempt_warning configuration
In order to the display a warn text on the last attempt
before the account is locked, we need update
config.paranoid to false as the devise documentation
explains, but currently config.paranoid is set to true,
so currently this feature is not working.

Adding "config.paranoid: false" implies further changes
to the code, so for now we updated the value
"config.last_attempt_warning = true" to false.