Load maximum_attemps and unlock_in from secrets
Currently the maximum attempts and unlock_in is being attempted to load from
the database through the Settings:

'''
(Setting["login_attempts_before_lock"] || 2).to_i
(Setting["time_to_unlock"] || 2).to_i.minutes
'''

The problem is that when you initialize these values in the devise initializer, you do
not yet have access to the database, therefore they are always returning the
hardcoded value, "2" in this case.

Now we make these values to be loaded through the secrets since we do have
access to them at the time the devise initializer is executed.

Update the default values with the values that we have in
production database:

Setting["time_to_unlock"]
 => "30"

Setting["login_attempts_before_lock"]
 => "5"

We also modify the code so that by default hours can be defined instead of
minutes (the same as the default value defined in the devise initializer).