Move settings configuration details to defaults method
Before this change, two important things depend on the format of each key,
where to render it in the administration panel and which kind of interface
to use for each setting. Following this strategy led us to a very complex
code, very difficult to maintain or modify. So, we do not want to depend
on the setting key structure anymore to decide how or where to render each
setting.

With this commit, we get rid of the key format-based rules. Each setting
has a Config object where we can explicitly define everything needed to
improve the administration settings.

Also, I removed most of the fake settings from tests and used real ones,
as the real ones already have a config set up. I only left the fake
settings for tests where we do not need access to the setting config.
Finding a comfy way to stub a config for fake settings is not worth the
effort required.