Avoid warnings during initialization
These warnings appear in the logs in the development environment. When
running the tests, they would appear in the standard error ouput if we
set `config.cache_classes = false` in the test environment but, since
that isn't the case, they don't.

To reproduce these warnings (or the lack of them), start a Rails console
in development and check the log/development.log file.