Remove visit_id from the debates table (updated)
This was added in commit 02f19aa4b, before we started tracking events.
I don't think we ever used it; in any case, we now use the `Ahoy::Chart`
class to deal with the stats Ahoy used to generate.

Updated: This commit has been updated. The original migration was
ActiveRecord::Migration[7.0], but actually in this project we are
use ActiveRecord::Migration[6.1]